Fix circular import between base_models and models/__init__.py
This error didn't happen when fluent_blogs.models was imported first.
This happens in the `runserver` as it performs a model scan,
but may not happen in production. This fix avoids the error taking
place.

Ideally, the either managers.py or base_models.py needs to be moved,
with a declaration warning included. This is out of the scope for now.

Closes #8